prov/efa: HMEM_CUDA support for emulated atomics
Adds support for emulated atomic calls to EFA for Cuda accelerators.
Atomics are only supported on machines that have GDR + RMA support.
This aligns with our overly restrictive FI_HMEM definition.
The first major change was switching every memcpy to be a ofi_hmem_*
memcpy. For fetch and compare operations, I added the result and
compare memory descriptors to rxr_atomic_ex. Lastly, I wrapped
the atomic operations on the target in ofi_hmem_memcpys with a temporary
bounce buffer in which the cpu only atomic operations are performed.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>