prov/sm2: Create inline freestack
Create an inline freestack which has 256 size sm2_xfer_entry (header +
user data in the 256 bytes). Initialize 4096 of these inline freestacks
per endpoint, which will expand the SMR footprint by 1.049 MB. This
freestack will be used in the inline protocol (implemented in future
commit) for sending very small messages. The advantage of the inline
protocol is it gives you better spatial locality, and allows you to fit
more things in the cache.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>