prov/sm2: Add Inline protocol
The inline protocol is a small message size optimization.  It uses a
seperate freestack of 256 byte xfer_entries in order to acheive better
spatial locality in the cache. From a performance perspective, the
cache efficiency is worth the extra algorithmic steps on the critical
send path. This optimization helps the most in large collective
operations.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>