prov/sm2: Use many shared memory files instead of 1 shared memory file
***** YOU NEED TO MANUALLY CHANGE LINE 456 of sm2_ep.c to the number of
ranks ******

Move each ranks region to a distinct shared memory file.  Keep the
common file to synchronize names, and then in the sm2_create() funciton
manually create a region with the processes gid from global file.  Put a
while loop to block until all files are mapped.  Needed to change the
queue a bit to support this as well.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>