prov/efa: Remove FI_THREAD_SAFE control interface
This patch improves the performance of the EFA provider by causing EFA
to not comply with the Libfabric API promise of offering thread safe
control interfaces.  When a user specifies FI_THREAD_DOMAIN, we will
make the control interfaces thread safe to the level of
FI_THREAD_DOMAIN. It is common for message endpoints to have a seperate
thread for just control calls, and those endpoints still want to claim
FI_THREAD_DOMAIN, instead of FI_THREAD_SAFE for performance reasons. The
EFA endpoint is an RDM endpoint, so the use case of a thread for the
control API is not an expected use case.  I created an issue for
adjusting the libfabric API
https://github.com/ofiwg/libfabric/issues/9665 to make RDM points not
maintaint the FI_THREAD_SAFE control interface promise. Those changes
might take a long time (if they ever) get passed the community. This
change gets the EFA Provider the performance boost immediately.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>