prov/efa: Do not memset rxe to 0 on init
Remove memset of struct efa_rdm_ope (size = 912 bytes) when we alloc an
rxe on a progress which progresses a message.  This reduces the time to
progress a message by ~40-55ns (3-5% on 1k msg sizes).  This commit
removes saftey guard rails in order to squeeze out additional performance
from the EFA provider.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>