prov/util: Make ep_list_lock noop for FI_THREAD_DOMAIN_RELAXED
This patch makes ep_list_lock a noop when user declared
FI_THREAD_DOMAIN_RELAXED.  The util_cntr and util_cq ep_list_locks are
used on the critical path for data transfer, so this change should
provide a performance improvement.  The util_av ep_list_lock should not
be used on the performance critical path, but it is changed for
simiplicty for the reader.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>